extract where transformer
